fix: move islands/shared helpers out of Blume's island scan root
Blume's discoverIslands recursively globs **/*.{jsx,svelte,tsx,vue}
under islands/ with no exclude mechanism, so it tried to treat
shared/{Button,Dropzone,FileIcon,FileList,IconButton,UploadFileList}.tsx
as top-level islands requiring a default export, failing the build
with [MISSING_EXPORT] errors since these are named-export helpers.

Moved islands/shared -> components/shared (outside the scan root) and
updated all 14 islands' import paths accordingly. Verified via
`blume build --isolated`.
