+++ apps/docs/islands/BasicExample.tsx
@@ -1,6 +1,6 @@
 import { useMediaDrop } from "react-mediadrop";
-import { Dropzone } from "./shared/Dropzone";
-import { FileList } from "./shared/FileList";
+import { Dropzone } from "../components/shared/Dropzone";
+import { FileList } from "../components/shared/FileList";
 
 export default function BasicExample() {
 	const { acceptedFiles, getRootProps, getInputProps, isDragActive } =
