+++ apps/docs/islands/FileDialogExample.tsx
@@ -1,7 +1,7 @@
 import { useMediaDrop } from "react-mediadrop";
-import { Button } from "./shared/Button";
-import { Dropzone } from "./shared/Dropzone";
-import { FileList } from "./shared/FileList";
+import { Button } from "../components/shared/Button";
+import { Dropzone } from "../components/shared/Dropzone";
+import { FileList } from "../components/shared/FileList";
 
 export default function FileDialogExample() {
 	const { acceptedFiles, getRootProps, getInputProps, open } = useMediaDrop({
